src/main/java/com/joe/models/Task.java
package com.joe.models;

import java.util.Date;

/**
 * Created by localadmin on 8/17/16.
 */
public class Task {

    private String name;
    private int id;
    private String category;
    private Date dateDue;
    private Date updatedAt;
    private Date createdAt;
    private Boolean isComplete;

    public Task() {
        this.createdAt = new Date();
        this.updatedAt = new Date();
    }

    public Date getUpdatedAt() {
        return updatedAt;
    }

    public void setUpdatedAt(Date updatedAt) {
        this.updatedAt = updatedAt;
    }

    public Date getCreatedAt() {
        return createdAt;
    }


    public void setCreatedAt(Date createdAt) {
        this.createdAt = createdAt;
    }

    public String getCategory() {
        return category;
    }

    public void setCategory(String category) {
        this.category = category;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public Task(int id) {
        this.id = id;
    }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public Date getDateDue() {
        return dateDue;
    }

    public void setDateDue(Date dateDue) {
        this.dateDue = dateDue;
    }

    public Boolean getIsComplete() {
        return isComplete;
    }

    public void setIsComplete(Boolean complete) {
        isComplete = complete;
    }

    protected void updateTime() {
        this.updatedAt = new Date();
    }
}
